% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgcricolors.R
\docType{package}
\name{jgcricolors}
\alias{jgcricolors}
\title{jgcricolors}
\description{
An R package to manage JGCRI colors for maps and charts.
}
\details{
\itemize{
\item Github: https://github.com/JGCRI/jgcricolors
\item Webpage: https://jgcri.github.io/jgcricolors}
}
