% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS49df}
\alias{mapUS49df}
\title{US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.}
\format{
A SpatialPolygonsDataFrame
}
\source{
Made with Natural Earth. \url{https://www.naturalearthdata.com/}
}
\usage{
mapUS49df
}
\description{
US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.
}
\keyword{datasets}
